pyproject.toml
[project]
name = "fraud-detection"
version = "0.1.0"
description = "Bank Transaction Fraud Detection System for Devoteam Interview"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Data processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",

    # Geospatial
    "geopandas>=0.14.0",
    "folium>=0.15.0",

    # Machine Learning
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "joblib>=1.3.0",

    # Deep Learning
    "torch>=2.2.0",
    "pytorch-lightning>=2.2.0",

    # Explainability
    "shap>=0.44.0",

    # MLOps
    "mlflow>=2.11.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Async & Caching
    "redis>=5.0.0",
    "httpx>=0.26.0",

    # Dashboard (optional - not needed for static CSV)
    # "streamlit>=1.31.0",

    # LLM (Vertex AI / Gemini)
    "google-generativeai>=0.5.0",
    "google-auth>=2.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "ipykernel>=6.29.0",
    "nbformat>=5.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
